bi-gcp-application/Data_Loading/requirements.txt
google-cloud-bigquery
pandas
pyarrow
google.cloud.logging
db-dtypes
google-cloud-bigquery-storage